+++ service/video/v1/client.go
@@ -6,8 +6,10 @@ import (
 	"github.com/RJPearson94/twilio-sdk-go/service/video/v1/composition"
 	"github.com/RJPearson94/twilio-sdk-go/service/video/v1/composition_hook"
 	"github.com/RJPearson94/twilio-sdk-go/service/video/v1/composition_hooks"
+	"github.com/RJPearson94/twilio-sdk-go/service/video/v1/composition_settings"
 	"github.com/RJPearson94/twilio-sdk-go/service/video/v1/compositions"
 	"github.com/RJPearson94/twilio-sdk-go/service/video/v1/recording"
+	"github.com/RJPearson94/twilio-sdk-go/service/video/v1/recording_settings"
 	"github.com/RJPearson94/twilio-sdk-go/service/video/v1/recordings"
 	"github.com/RJPearson94/twilio-sdk-go/service/video/v1/room"
 	"github.com/RJPearson94/twilio-sdk-go/service/video/v1/rooms"
@@ -19,14 +21,16 @@ import (
 type Video struct {
 	client *client.Client
 
-	Composition      func(string) *composition.Client
-	CompositionHook  func(string) *composition_hook.Client
-	CompositionHooks *composition_hooks.Client
-	Compositions     *compositions.Client
-	Recording        func(string) *recording.Client
-	Recordings       *recordings.Client
-	Room             func(string) *room.Client
-	Rooms            *rooms.Client
+	Composition         func(string) *composition.Client
+	CompositionHook     func(string) *composition_hook.Client
+	CompositionHooks    *composition_hooks.Client
+	CompositionSettings func() *composition_settings.Client
+	Compositions        *compositions.Client
+	Recording           func(string) *recording.Client
+	RecordingSettings   func() *recording_settings.Client
+	Recordings          *recordings.Client
+	Room                func(string) *room.Client
+	Rooms               *rooms.Client
 }
 
 // NewWithClient creates a new instance of the client with a HTTP client
@@ -44,14 +48,16 @@ func NewWithClient(client *client.Client) *Video {
 				Sid: compositionHookSid,
 			})
 		},
-		CompositionHooks: composition_hooks.New(client),
-		Compositions:     compositions.New(client),
+		CompositionHooks:    composition_hooks.New(client),
+		CompositionSettings: func() *composition_settings.Client { return composition_settings.New(client) },
+		Compositions:        compositions.New(client),
 		Recording: func(recordingSid string) *recording.Client {
 			return recording.New(client, recording.ClientProperties{
 				Sid: recordingSid,
 			})
 		},
-		Recordings: recordings.New(client),
+		RecordingSettings: func() *recording_settings.Client { return recording_settings.New(client) },
+		Recordings:        recordings.New(client),
 		Room: func(roomSid string) *room.Client {
 			return room.New(client, room.ClientProperties{
 				Sid: roomSid,